unsuccessful Deploy, closes #5, closes #4, closes #2, closes #1, closes #6

+++ app/resources/destory-recipe.tsx
@@ -1,15 +1,45 @@
-import { redirect } from "react-router";
+import { data, redirect } from "react-router";
 import { prisma } from "~/lib/db.server";
 import { requireAuthSession } from "~/lib/session.server";
+import {
+  formatUserShoppingList,
+  updateUserShoppingList,
+} from "~/lib/shopping-list";
 import type { Route } from "./+types/destory-recipe";
 
 export async function action({ request, params }: Route.ActionArgs) {
   const session = await requireAuthSession(request);
 
-  await prisma.recipe.delete({
+  const recipe = await prisma.recipe.findUnique({
     select: { id: true },
     where: { id: params.recipeId, userId: session.user.id },
   });
+  if (!recipe) {
+    throw data("No recipe found", { status: 404 });
+  }
+
+  await prisma.recipe.delete({
+    select: { id: true },
+    where: { id: params.recipeId },
+  });
+
+  const user = await prisma.user.findUniqueOrThrow({
+    select: { shoppingList: true },
+    where: { id: session.user.id },
+  });
+  const shoppingList = formatUserShoppingList(user.shoppingList);
+
+  const isRecipeInShoppingList = shoppingList.includes(recipe.id);
+  if (isRecipeInShoppingList) {
+    // Remove the deleted recipe id from the shopping list
+    const nextShoppingList = updateUserShoppingList(shoppingList, recipe.id);
+
+    await prisma.user.update({
+      select: { id: true },
+      data: { shoppingList: nextShoppingList },
+      where: { id: session.user.id },
+    });
+  }
 
   return redirect("/recipes");
 }
